what ive done
so i went to bootstrap.com and went to examples and chose the "blog" sample. i viewed the source code and coppied it to my index file. 
from there i went to the functions and enqueue the style and js scripts and made sure to add action to each. i then split my index file into
header and footer sectons.
i made a style sheet just for the wordpress info and keept their blog.css in the css file as they had theirs.
when i try to view the page in wp i see all the script with absoultely no style elements what so ever. i have followed with the treehouse
course and have been able to get the treehouse to wp project to load and have added sidebars and different pages ect but when i do the same 
here it doesnt work.